feat: refactor built-in AI structure and add global download progress tracking
