+++ src/shared/built-in-ai/internal/lifecycle/store.ts
@@ -5,9 +5,10 @@ import {
   UnavailableError,
   UnsupportedError,
 } from "../../errors.ts";
-import { abortError, mergeSignals, raceAbort } from "../../util/signal.ts";
+import type { Status } from "../../types.ts";
+import { abortError, mergeSignals, raceAbort } from "../signal.ts";
 import { createInstance } from "./create-instance.ts";
-import type { AINamespace, DestroyableInstance, Status } from "./types.ts";
+import type { AINamespace, DestroyableInstance } from "./types.ts";
 
 function hasUserActivation(): boolean {
   return navigator.userActivation?.isActive ?? false;
