+++ src/shared/built-in-ai/internal/lifecycle/useLifecycle.test.tsx
@@ -16,8 +16,8 @@ import {
   UnavailableError,
   UnsupportedError,
 } from "../../errors.ts";
-import { makeAIFake } from "../../test-fakes/ai-namespace-fake.ts";
-import { snapshotProgressFor } from "../progress/progress-store.ts";
+import { snapshotProgressFor } from "../progress-store.ts";
+import { makeAIFake } from "../testing/ai-namespace-fake.ts";
 import { useLifecycle } from "./useLifecycle.ts";
 
 interface TestOptions {
