+++ src/shared/built-in-ai/internal/progress-store.test.ts
