+++ src/shared/built-in-ai/internal/progress-store.ts
