+++ src/shared/built-in-ai/internal/signal.test.ts
