+++ src/shared/built-in-ai/internal/signal.ts
