+++ src/shared/built-in-ai/internal/stream.test.ts
