+++ src/shared/built-in-ai/internal/stream.ts
