+++ src/shared/built-in-ai/internal/testing/ai-namespace-fake.ts
