+++ src/shared/built-in-ai/internal/testing/instance-fakes.ts
