+++ src/shared/built-in-ai/proofreader/useProofreader.test.tsx
@@ -1,7 +1,7 @@
 import { afterEach, describe, expect, test, vi } from "vitest";
 import { renderHook } from "vitest-browser-react";
-import { makeAIFake } from "../test-fakes/ai-namespace-fake.ts";
-import { buildProofreaderInstance } from "../test-fakes/instance-fakes.ts";
+import { makeAIFake } from "../internal/testing/ai-namespace-fake.ts";
+import { buildProofreaderInstance } from "../internal/testing/instance-fakes.ts";
 import { useProofreader } from "./useProofreader.ts";
 
 afterEach(() => {
