+++ src/shared/built-in-ai/proofreader/useProofreader.ts
@@ -1,6 +1,6 @@
 import { useState } from "react";
-import type { BaseHookReturn } from "../internal/lifecycle/types.ts";
 import { useLifecycle } from "../internal/lifecycle/useLifecycle.ts";
+import type { BaseHookReturn } from "../types.ts";
 
 /**
  * Options for {@link useProofreader}. Mirrors `Proofreader.create()`. Compared
