+++ src/shared/built-in-ai/translator/useTranslator.test.tsx
@@ -1,7 +1,7 @@
 import { afterEach, describe, expect, test, vi } from "vitest";
 import { renderHook } from "vitest-browser-react";
-import { makeAIFake } from "../test-fakes/ai-namespace-fake.ts";
-import { buildTranslatorInstance } from "../test-fakes/instance-fakes.ts";
+import { makeAIFake } from "../internal/testing/ai-namespace-fake.ts";
+import { buildTranslatorInstance } from "../internal/testing/instance-fakes.ts";
 import { useTranslator } from "./useTranslator.ts";
 
 afterEach(() => {
